requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
exceptiongroup==1.2.1
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
python-decouple==3.8
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
